+++ packages/eyes-webdriverio-5/scripts/start-chromedriver.js
@@ -4,3 +4,4 @@ if (!process.env.SKIP_CHROMEDRIVER || process.env.CVG_TEST_REMOTE) {
   const returnPromise = true
   chromedriver.start(['--port=4444', '--url-base=wd/hub', '--silent'], returnPromise)
 }
+process.exit(0)