+++ packages/visual-grid-client/src/sdk/checkWindow.js
@@ -56,8 +56,6 @@ function makeCheckWindow({
     useDom,
     enablePatterns,
     ignoreDisplacements,
-    source,
-    referrer,
   }) {
     if (target === 'window' && !fully) {
       sizeMode = 'viewport'
@@ -66,7 +64,7 @@ function makeCheckWindow({
     } else if (target === 'region' && region) {
       sizeMode = 'region'
     }
-    fetchHeaders['Referer'] = referrer
+    fetchHeaders['Referer'] = url
 
     const accErr = isInvalidAccessibility(accessibility)
     if (accErr) {
@@ -262,7 +260,7 @@ function makeCheckWindow({
         domUrl: domLocation,
         checkSettings,
         imageLocation,
-        source,
+        url,
       }
 
       return !isSingleWindow ? wrapper.checkWindow(checkArgs) : wrapper.testWindow(checkArgs)