+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## Unreleased
 
+- updated VGC for using source instead of referer
+- updated to @applitools/eyes-sdk-core@10.1.2
+- updated to @applitools/visual-grid-client@14.0.0
 
 ## 3.10.4 - 2020/5/7
 
 - updated to @applitools/dom-snapshot@3.5.2
 - updated to @applitools/eyes-sdk-core@10.0.1
-- updated to @applitools/visual-grid-client@13.7.6
 - updated to @applitools/visual-grid-client@13.7.7
 
  ## 3.10.3 - 2020/4/30